wp-backdoor-user.php
Hi, 

I've made several enhancements on the file I'd like to propose, here is a list:
1) line 66 : add a boolean variable to make the file deleting optional
2) line 69 : reinforce security by accessing the file with a GET parameter which is the hash of the file (so as only uploader can know it)
3) line 83 : include 'user.php' because deleting a user throw the error that 'wp_user_delete' is unknown function (also buggy in your version)
4) line 123 : optional delete of the file with variable set previously
5) line 148 : same
6) line 215 : within this optional file deleting, create HTML content warning to be displayed in the page
7) line 237 : complete rewrite of the HTML part, including Font Awesome & Bootstrap 4 for a fancy display

I didn't touch too much of your functional script as it was OK, but I'm a webdesigner and I was a kinda sad of the presentation so I decided to dive into it. Please don't be upset about that.

The script certainly need to be improve but this is enough for today. I suggest to increment the version from 2.1 to 3.0.

I'm waiting forward your thoughts about it, feel free to contact me and give your impressions.
Thanks in advance,
Geekly
Fanch